fix: Cleanup unused functionality in remote code
Previously, when a user used --remote, it'd clone the folder
into a tmp folder without changing the cwd and scan it there. This was
fragile and useless beyond making the command a little cleaner. The
cleanliness to the user was not worth the ugliness to our code. So now
--remote will just clone the folder into the cwd (assuming it's empty).
Now we don't do some weird conditional cwd change either (sorry @kopecs
and @mjambon, yall were right).

Downside is that users have to cleanup after themselves but oh well