test: add semgrep-core-proprietary binary gha artifact cache (#9880)
For some kinds of changes, we need the pro binary, so we can make sure
that pytests work correctly in pro mode.

Example PR that would benefit from this change: [enabling interproc
taint scans by default](https://github.com/semgrep/semgrep/pull/9872) or
[adding pro flags to the metrics
test](https://github.com/semgrep/semgrep/pull/9258).

PRs that would not benefit from this change are PRs in semgrep that
require changes from Pro. The workflow brings in the develop branch, so
WIP in pro would not be brought in.

Eventually, we are headed towards a monorepo. If that is coming soon,
then this PR can probably be ignored.

This PR extends the `check-semgrep-pro` workflow to also upload the
binaries (both semgrep-core and semgrep-core-proprietary) to gh cache.
The semgrep-core-proprietary is built using code from develop and
includes the most recent changes here.

I needed to change the container to alpine to force a statically linked
binary. I haven't found a way to statically link on ubuntu yet. We need
a statically linked binary so that the tests don't need to depend on the
tree sitter library being installed. (This is also the case with
semgrep-core that is built in alpine too.)

When downloading the binary, also needed to make sure
`pro-installed-by.txt` has the correct info and is in the same directory
as the binary. Otherwise, the CLI won't use the binary.

Added a test `test_taint_intrafile.py` which uses the `--pro_intrafile`
flag and the test passes. No other tests requiresthe pro binary.