fix(secrets): validator rules should only respect validation state actions (#9943)
## Context

If a validator rule was set to blocking, every validation state will be
blocked. For example, if we have an invalid finding and the rule is set
to block for valid findings, then the invalid finding will block too.

## Solution

If a finding has a validation state, then only respect actions in
dev.semgrep.validation_state.actions. We made the mistake of taking
`dev.semgrep.actions` into account.

---------

Co-authored-by: Cooper Pierce <cooper@semgrep.com>