refactor/fix: Perform autofix application in OCaml (#9961)
This introduces an RPC mechanism by which we can call OCaml code in
semgrep-core from Python. This will allow us greater flexibility when
incrementally migrating functionality to OCaml as part of the osemgrep
project.

As a motivating example, I have moved autofix application to OCaml.
Brandon attempted this in #9229, but was forced to leave some autofix
functionality in Python. After running the initial scan in
`semgrep-core`, we run a baseline scan in certain cases. We can't do
that if the first call to `semgrep-core` has already applied autofixes,
dirtying the repository. Therefore, we need to be able to do it outside
of the main call to `semgrep-core`.

I've prepared for this change by adding additional autofix e2e tests in
#9948, #9914, and #9857.

This is labeled a "refactor/fix" instead of just a refactor because this
actually fixes an issue with an autofix application on a line that
contains a multi-byte character (in UTF-8 encoding), where the autofix
occurs after the multi-byte character. See the changed e2e test
snapshots.

For more details see the README.txt file.

Test plan: Automated tests, plus manual testing of many (but not all) of
the error conditions I handle related to RPC calls.