Add semicolon token for variable definition in AST_generic
This should also fix autofix on var defs in a few languages.
The motivation for this PR is romain's stubber
https://github.com/semgrep/semgrep-proprietary/pull/1397
which currently don't work well because the range of vardef
is not correct and so can't really be removed because the semicolon
is not present in the AST (CST really).

A different PR will fix the issue for Java.
This will help SAF-928

test plan:
make
make test