feat: interproc taint scans run by default for logged in users (#9872)
Previously, we required the Pro engine toggle to be on to run Pro scans,
including interproc taint (PRO_INTRAFILE) scans. Now that we've
validated that these PRO_INTRAFILE scans give good improvement in
results with virtually no perf regression, and that we have the pro
binary in the docker image, it's time to give users more powerful
analysis by default.

Test plan: automated unit tests, canary image

Notes:
- With the current PR, we no longer run `semgrep ci` tests with OSS
analysis. IMO this is fine; there are likely to be few differences and
we would like to get to a world where nobody runs `semgrep ci` with our
OSS analysis
- A bigger problem is that `semgrep ci` tests won't run if the pro build
is broken, which happens from time to time. I'm ok with this as long as
we plan to merge the repos soon-ish
- TODO: I'd like to move the handful of python tests in
semgrep-proprietary to this repo and run these pytests in
semgrep-proprietary as well