ci: add platform dep installation to ocaml5
We think this may be necessary based off of some recent CI failures
related to libev not being able to be linked when git-unix is being
used. See <https://github.com/semgrep/semgrep/actions/runs/8474645693/job/23221419153>.

Not clear why this passes for some PRs and not for others, but we think
it may be related to a dirty state on the container. It may be better to
have the `install-deps-for-semgrep-core` to have a prereq on or
otherwise run `install-deps-$PLATFORM-for-semgrep-core`, but we haven't
evaluated this deeply.