fix: Correct match-based-id for lockfile-only SCA findings (#10020)
https://www.notion.so/semgrep/Fixing-SSC-match-based-id-and-triage-state-d4676bfa62a2439eb026ac22b3d6c119

Given a lockfile with two instances of a library, say `foo@1.0.0` and
`foo@1.0.1`, if both are matched by a lockfile-only SCA rule, semgrep
will produce two findings, but these findings will incorrectly have
identical `match_based_id`s, so Semgrep App considers them identical
findings. This happened because lockfile-only findings were created
without the use of the `RuleMatches` class, which handles updating the
`match_based_id` of each finding. We actually can't use `RuleMatches`,
because it ends up changing the whole `match_based_id`, but we only want
to change the final index, to avoid duplicating findings in the
database, so we re-implement this functionality manually.

In the case where the lockfile contains two instances of a library at
the _same_ version, then we will still produce two separate findings
with different `match_based_id`s. This is technically incorrect, but all
of our code is built assuming this is possible, so it can be corrected
in future PRs.

Due basically to luck, for a given set of lockfile-only findings from a
single rule, the Issue with `match_based_id` ending with index 0 in the
database will be the same before and after this change. This is because
before upload to the database these findings are put into a dictionary
keyed on their `match_based_id`, and the findings are processed in
reverse order, so the dependency that appeared _first_ in a lockfile
gets put into this dictionary _last_. Before the change, when these
findings all had the same `match_based_id`, the first matched dependency
in the lockfile becomes the only thing in the database. After the
change, the first thing in the lockfile still gets index 0, since it
appears first. This means the single issue that gets selected as the
representative of the vuln group for these issues won't change, since we
select the oldest issue, and the one ending in index 0 existed
previously, so it's the oldest.

---------

Co-authored-by: Clara McCreery <chmccreery@gmail.com>