refactor(osemgrep): move logic outside of sarif formatting code (#10028)
There is logic in sarif formatting code that checks if
1. a user is logged in and
2. the scan is using a registry

Having these states be determined in the formatting code makes it
difficult to isolate this function on its own. Isolation helps here
because we'd like to call formatting code through an RPC for osemgrep
work, where the ocaml function's behavior depends solely on its
parameters.

This PR moves some of this logic outside of sarif formatting code.

This is osemgrep code only, which is not executed by default. It is
guarded by the `--experimental` flag.

Test plan: ran `make osempass` which ensures previously passing osemgrep
tests are still passing.