Add two test cases for overlapping matches in spacegrep. (#10038)
One of them behaves intuitively (`a ... b`) but not the other (`a ...`).
I marked the second one as xfail.

Given a target file containing `a a b`:
* `a ... b` is expected to match only `a b` and works as expected;
* but `a ...` is expected to match `a a b` based on naive expectations
or common intuition. It doesn't work.

I suggest we merge this branch and have the bug report link to the tests
added by this PR. It will simplify the work of actually fixing it.