ci: split osx-setup-for-release
This is so we can avoid issues in pro's CI due to additional opam
dependencies we would ideally like to install after what is now
`osx-setup-opam-for-release` but before `osx-setup-brew-for-release`.
The reason for this is so that we can remove pcre2 to force static
linking, but only after `opam install`, since `opam` relies on `git`,
which is built with a (dynamically linked) dependency on pcre2.