fix: do not run historical scans on diff scans (#10051)
In some instances, the App could send that historical scanning was
enabled to a diff scan. This ensures that in that event, we don't run
the historical scan (for time reasons) and print a diagnostic.

In the future, when we have a historical diff scan, that logic can be
inserted here instead.