revert: undo libpcre2 migration so we can fix the build first (#10064)
The pro build is failing in osx. We try to link pcre2 statically in
semgrep, and to do so, we remove the dynamic library from the system.

However, in the osx build, opam requires git, which requires the
dynamically linked pcre2 library. But we removed it because we want to
statically build semgrep.

While we figure this out, we've decided to revert the migration.