refactor(lsp): use search env, clean up (#9930)
This is a stacked diff whose parent is #9922.

## What:
This PR used to make it such that we use `Match_search_mode` instead of
calling `run_semgrep`, but I no longer think that's a good idea. So now
it just rearranges some stuff.

## Why:
We used to use `Match_search_mode` because I thought it would save some
work, and also because I wanted to leverage the match hook to do some
streaming. However, the PRs after this won't use the match hook at all,
instead opting to do a per-file scan basis. So we no longer need to do
that.

## Test plan:
Tested manually.

Closes CDX-278