chore: pin tls-mirage version (#10069)
Windows doesn't seem to like Tls_mirage 0.17.4. Example failing workflow
from last week:
https://github.com/semgrep/semgrep/actions/runs/8573190639/job/23499854171?pr=10064#step:8:906

So we need to pin the tls-mirage version to 0.17.3 for now.

I think the dependency is git-unix -> git-mirage -> git-paf -> paf ->
tls-mirage. It seems like we use the most recent version of git-unix. I
don't really know why paf 0.3 is being picked up, when 0.5 is the latest
version.

There's a more recent commit in paf:
https://github.com/dinosaure/paf-le-chien/commit/1751aec7161c5c449c887910c48fcaf526bddfb9#diff-69ea6088051956fa16649edb6c1460d9e6ccdb49778943abff2a5e127ec2b412R16,
and it seems like it may work with tls-mirage 0.17.4, so we should
probably re-visit later.

That said, this is blocking our current release, and I think pinning the
version of tls-mirage so it can build in Windows is a good short term
solution for now.