chore(lsp): dont search python thrice (#9941)
This is a stacked diff whose parent is #9940 

## What:
Our approach for inferring languages considers `Python`, `Python2`, and
`Python3` separately. This means we scan them all, essentially running
Python scans three times.

## How:
If we see that`Python3` is a valid language, we will just scan with t,
and get rid of `Python` and `Python2`.