debug: Do not log any debug message unless it matches some tag
If no debug tags are specified, we print nothing. We don't want --debug
output to be enourmous, that tends not to be useful. You should instead
first consider what exact debug info you need (i.e. what "tags"). This
is also a perf problem for the Python CLI that captures all this output
in-memory (despite it shouldn't do that in the first place...).

test plan:
Run semgrep-core with --debug