Remove CODEOWNERS for .github
Recently emma and cooper had issues releasing semgrep
because I was OOO and they could not modify
some GHA workflows. Simpler to not have any codeowners
for this critical part of the semgrep repo.

test plan:
see if I'm assigned on this PR since it modifies a file
in .github/