feat(lsp): match previews for `/semgrep/search` (#9949)
This is a stacked diff whose parent is #9945 

## What:
This PR generates the match previews for the VS Code extension via the
LSP, instead of grabbing the text later in the extension.

## Why:
Each time we touch the files in the extension, this generates a
`onTextDocumentOpen`, which the LSP then needs to handle. These can
build up and cause slowdowns.

## How:
We implement a "best effort" preview showing logic, which tries to find
the nearest "natural" place to start the match (currently a space). If
there is not such a location which leaves the match still acceptably
visible, we just intercept in the middle of some phrase.

## Test plan:
<img width="458" alt="image"
src="https://github.com/semgrep/semgrep/assets/49291449/86b30b9c-50eb-4a8f-a4df-56370fbb1014">