Start using Logs.src instead of tags
We're currently abusing tags for something it was not designed for.
The standard convention in the OCaml libraries world
is to use the Logs "src" to separate the logging
done by different libraries, so let's do the same
in the semgrep codebase.
See https://erratique.ch/software/logs/doc/Logs/index.html#srcs

test plan:
```
$ make check
...
[00.04][WARNING]: Support for Jsonnet rules is experimental and currently meant for internal use only. The syntax may change or be removed at any point.
[00.92][ERROR](commons.pcre): PCRE error: BadUTF8 on input "\127ELF\002\001\001\003\000\000\000\000\000\000\000\000\003\000>\000\001\000\000\0000\021\200\001\000\000\000\00
...
```