feat(lsp): multiple pattern search (#9979)
This is a stacked diff whose parent is
https://github.com/semgrep/semgrep/pull/9949

## What:
This PR makes it so that the search commands accept a _list_ of
patterns, each annotated with a Boolean denoting their "positivity",
rather than a single pattern. This effectively simulates "simple mode",
but via the LSP interface.

## Why:
Searching with a single pattern is reasonably limited. You can only
search for things expressible in a single pattern, and you cannot filter
anything out. This makes it rather difficult to use effectively,
especially in the context of mass refactor, when you want to make sure
you are only matching things which are relevant to you.

## Test plan: 
Tested manually.