Manually link on MacOS to allow for static linkage
In general, we want to statically link to make it easier to distribute
binaries since they won't carry runtime dependencies on various
libraries which would need to be separately installed or managed.
Additionally, this can be a bit fragile depending on which paths the
loader uses, and where dylibs are expected to be (and how possible
differences in the CI build versus other environments).

However, on MacOS, it is a bit difficult to ensure libraries are linked
statically. It is not possible to link all libraries statically.
Notably, libSystem cannot (officially) be statically linked against,
since MacOS does not have a stable syscall ABI, but rather relies on
libSystem being stable. See, e.g.,
- <https://developer.apple.com/forums/thread/706419>
- <https://developer.apple.com/library/archive/qa/qa1118/_index.html>
- `man 1 ld`

As a result, we cannot use a similar strategy to our Alpine build (which
works for Linux, since it is a rare case with a stable syscall ABI, See
<https://unix.stackexchange.com/questions/473137/do-other-unix-like-kernels-have-stable-syscall-abis>).

Previously we relied on a workaround where we deleted the `.dylib` versions
of libraries, leaving only the `.a` versions, thus forcing them to be
linked statically. However, this breaks the build in the case where
later build steps require programs which dynamically linked against
those libraries. Notably, this occurs for PCRE2. See
<https://github.com/semgrep/semgrep/pull/10064>.

One might hope that we could specify the linker to use the `.a` version
without deleting the `.dylib`, but the only way to do this is to disable
autolinking and specify each library individually. Otherwise, on MacOS,
if both are present, the linker will link dynamically regardless.

Here we change the process on MacOS to manually specify each library,
and augment our existing scripts to generate the appropriate list of
flags. This does require marginally more work when adding a library
(since the stubs or library may need to be added to the list), but it
helps to close the gap between what is run locally and in CI, and local
builds will fail in the same way, so debugging is improved.