feat(osemgrep): support SARIF formatting in osemgrep (#10113)
SARIF formatting in ocaml has already been implemented for a while. This
PR hooks it up with RPC, so it can be called from python code and be
incrementally rolled out to users sooner.

The user needs to specify `--sarif` along with
`--use-osemgrep-format-output` to use ocaml's implementation of SARIF
formatting.

Added a new formatter to the list of available formatters. The
implementation is in osemgrep_sarif.py, which calls the RPC and compares
its result to the python version. If the results are different, fallback
to python output and log at verbose level.

In terms of correctness, users will never see a json whose content is
different than pysemgrep. (The formatting may be different because ocaml
doesn't indent or add new lines, but the json content is never
different.)

The implementation for osemgrep_sarif.py and rpc.ml may look bulky, but
it's mostly marshaling arguments to pass to an existing function. I
think this is still a good step towards removing python code.

OutputSettings takes in a set of output formats that a user would like
to output with osemgrep. Currently, we only support 1 output at a time
so this might not look like it makes sense, but there's also another PR
in flight that supports multiple outputs
(https://github.com/semgrep/semgrep/pull/10112), and this osemgrep
formatting will need to be eventually incorporated.

Performance numbers:
* for a large payload (~3000 rules and ~2000 findings)
  * I ran `semgrep scan --config r/all semgrep-repo`
  * ~92 seconds total time
  * ~2 seconds e2e RPC call
  * ~45ms ocaml formatting
  * ~144ms python formatting
  * ~280ms validating whether osemgrep == pysemgrep or not
* for a small payload (1 rule and 1 finding)
* I ran `semgrep scan --config
r/python.lang.security.audit.dangerous-os-exec-tainted-env-args.dangerous-os-exec-tainted-env-args
`semgrep-repo/cli/bin/semgrep`
  * ~1 second total time
  * ~58ms e2e rpc call
  * <1ms ocaml formatting
  * <1ms python formatting
  * <1ms validating whether osemgrep == pysemgrep or not

I think this performance hit is acceptable, while it allows us to
measure how safe the rollout is going.

Tested with make e2e to ensure no regressions. Added a test for
osemgrep. It is skipped now, because there is a mismatch in the output.
The mismatch in the test has to do with sorting the rules, but the rest
of the payload, including the matched based id is the same. The fix will
be in a separate PR. This PR focuses on having the first version of
SARIF rpc.

Fixes SAF-978