fix: increase "mark scan completed" timeout from 20 to 30 minutes
When semgrep ci finishes scanning, it uploads the findings to the app,
then marks the scan as complete.

The "mark complete" part is a polling loop. The CLI will successfully
return when the app finishes processing the findings, or retry if the
app is not finished yet.

This polling loop has a 20-minute timeout. This PR increases it to 30
minutes, since we're starting to see some scans that take more than 20
minutes to mark complete.

This could also be done with an environment variable, but the team may
need more discussion before introducing even more knobs to the CLI, so
this PR just increases the hardcoded value.