ci: change root for flags.sh
Currently when semgrep is included in semgrep-proprietary as a
submodule, %{workspace_root} expands to the repository root of
semgrep-proprietary, not semgrep. This means the incorrect script is
used when attempting to build the targets from semgrep. Instead, we
should use %{project_root}, which should consistently be the semgrep
repository root.