ci: ensure make exits with 2 when build fails
As a result of the pipeline we introduced to save the output of dune so
we could provide a more helpful message when a linking error occurs,
make's exit status is now (incorrectly) always 0. This reworks the
pipeline to run with `-o pipefail` so if dune fails, the exit status is
correct (i.e., 2).