Update returntocorp refs, mainly for docs automations (#10137)
We still have a bunch of returntocorp references in `semgrep/semgrep`,
which is generally OK due to the GitHub redirect, but there are a few of
these that appear in help outputs and such which are used in places that
are already updated (like the docs site), and end up looking funny.

This PR updates a few of the more problematic cases to use
semgrep/semgrep. It was going to be rather of a Project to do all of
them so I focused on ones that seemed the most useful to fix.