Update Testo, fixing temporary path masking bug (#10114)
The bug was causing e.g. `/var/tmp` to be replaced by `/var<TMP>` if
`/tmp` was the temporary folder path. See
https://github.com/semgrep/testo/pull/56

This is why we had to mask the longer path first (`/a/b/c`) and only
then the shorter path (`/b/c`) that happened to be a substring of the
former.