ci: use pkg-config for all possible deps (#10141)
Changes pcre and pcre2 to use pkg-config (instead of brew prefix) for
portability. This should make the brew formula easier to maintain.

Libev does not support pkg-config. See, e.g.,
<https://www.mail-archive.com/libev@lists.schmorp.de/msg02088.html>,
<http://lists.schmorp.de/pipermail/libev/2024q1/002940.html>. As a
result we still use the brew prefix, but with the option of an
environment variable for the build to override the location.