fix: send notification when rules have refreshed
We now send a notification when rules have refreshed so we can do things
like start tests on the IDE side. And so we can restart settings and
then run a full workspace scan.

Also we now use parmap to get the # of cores for the default number of
jobs which should speed things up for people by default