fix(secrets): move `--historical-secrets` to pro options
The `--historical-secrets` flag was in the "Output format" option group
in error, which led to issues as pysemgrep treated it as mutually
exclusive with output options like `--json`. This moves it to the
correct group.

Test plan: existing tests for regression. Will add a test in pro to
ensure that usage with `--json` works as expected.