const-prop: Recognize Python's "..." * N expressions as constant (#10155)
In Python, `"a" * 3` evaluates to `"aaa"`. We will not evaluate these
expressions but we will at least recognize them as (arbitrary) string
constants so they can be matched by `"..."` patterns. We don't check
whether the N is indeed an integer, but this should be fine.

Closes FLOW-75

test plan:
make test # two new tests