test(osemgrep): make osemgrep sarif format test pass (#10146)
For the current e2e we have, the only mismatch between pysemgrep vs
osemgrep is the ordering of rules

The ordering of rules in the sarif output format shouldn't matter to the
users, so there's no code update here. Only tests that compare the json
objects.

Tested with `make e2e`.