fix(osemgrep): SARIF correctly handles CWE tags (#10161)
When CWE metadata is present in the rule, the "security" tag should
automatically be added to the SARIF output format. Moreover, tags should
be unique. This PR updates osemgrep to do this.

Also added debug logging, guarded by `--debug`, which prints out the
diff when osemgrep and pysemgrep don't match. This was useful for
testing and finding out this bug.

Tested with `make e2e`.

Fixes SAF-991.