fix: support callv and proxies for lsp.js
Previously we added support for proxies in #9994. But it broke LSP.js
networking as we were forced to use callv to support proxies as a
workaround, and callv isn't supported in cohttp-jsoo!. So we have to
patch callv and patch our XHR js library to enable proxies for this
stuff. I'm going to try and make a bunch of PRs to cohttp and ca-certs
etc to fix this upstream, but for now we have this fix