Update Testo (#10173)
This adds a `name` file containing the name of the test to each snapshot
folder so that we can review dead snapshots. See
https://github.com/semgrep/testo/pull/66

This update also includes a function `Testo.with_capture` for checkout
stdout or stderr against a string (rather than a file).

It also includes this last-minute change in Testo which allows sharing
tags (type and tag registry) between the Testo and Testo_lwt libraries
(https://github.com/semgrep/testo/pull/68).