fix: support callv and proxies for lsp.js (#10172)
Previously we added support for proxies in #9994. But it broke LSP.js
networking as we were forced to use callv to support proxies as a
workaround, and callv isn't supported in cohttp-jsoo!. So we have to
patch callv and patch our XHR js library to enable proxies for this
stuff. I'm going to try and make a bunch of PRs to cohttp and ca-certs
etc to fix this upstream, but for now we have this fix.

## Test plan

### For testing js w/out proxy
```bash
make build-semgrep-jsoo-debug
make -C js build
make -C js test
```
### For testing JS + proxy
```bash
make build-semgrep-jsoo-debug
make -C js build
make -C js/language_server
# copy js/language_server/dist into extension
# Run vscode with proxy and Use JS to confirm it works
```

TODO add MIT license to VSCode