chore(js): dont print raised ocaml exns in js (#10183)
## What:
This PR makes it so that we no longer log the name of OCaml exceptions
raised in JSCaml code.

## Why:
This causes some degree of log spew. We already raise the exception, we
don't need to unconditionally print the name of the exception.

## How:
Just removed the print.

## Test plan:
Tried it locally, and it no longer prints when a `parsePattern` raises a
parsing error.