fix: allow pattern matching on go struct tags (#10179)
This fix closes
[SAF-949](https://linear.app/semgrep/issue/SAF-949/go-support-struct-json-tags)
by allowing users to pattern match on the contents of Go struct field
tags. There are a couple of test cases added. One of which shows how to
match on tags without regards to the order of the string.

# test plan
Added some test cases. CI should continue to work.

---------

Co-authored-by: Nat Mote <nat@semgrep.com>