Fix dockerfile parsing bug due to missing newline (#10185)
This isn't a fix in tree-sitter-dockerfile as it should be but it's the
same workaround we had in place for parsing patterns. Dockerfile
patterns often lack a trailing newline, but regular dockerfiles may also
lack a trailing newline. This PR adds the missing newline to the input
if it's missing. Such errors where tree-sitter inserts a token were
silently ignored in semgrep as well as in our parsing tests. I added a
field `tolerated_errors` to the parsing result that allows our tests to
be strict about this. We now have two tests that would fail without the
newline fix. Semgrep still ignores "missing token" errors (saves falling
back to a slower pfff parser, as already annotated in the code).

There is a deeper issue reported by
https://github.com/semgrep/semgrep/issues/10084 where their input (our
new test file `no-trailing-newline-segfault.dockerfile`) causes a
segmentation fault. It is reported incorrectly as an out-of-memory error
by the Semgrep Playground but this is not very important. Adding the new
line at the end of the input allows for proper parsing without a
segfault but this should be investigated further to determine if it's a
bug in tree-sitter-dockerfile's `scanner.c`, in the OCaml/tree-sitter
bindings, or something else.

Update: the issue is known and tracked at
https://github.com/camdencheek/tree-sitter-dockerfile/issues/22. It
should be fixed with a tree-sitter upgrade. (thx @kopecs)

Fixes https://github.com/semgrep/semgrep/issues/10084

semgrep-pro: https://github.com/semgrep/semgrep-proprietary/pull/1469