refactor: Include `name` in XML kind
This is a followup to #10182, where I added `id_info` to the XML kind.

I added a TODO to consider changing this to `name`, since in JS/TS,
`<Foo.Bar />` is legal, so we should be able to use `IdQualified` in
such cases.

I figured we could just change it when we need it. However, I failed to
understand that the Pro Engine traverses the AST looking for `name`s in
order to compute taint signatures in dependency order. This led to
strange problems with my work in progress. Instead of updating that
traversal to special-case XML kind, I'm just putting a `name` here like
I eventually intended to do anyway.

Test plan: Automated tests