Remove use of returntocorp/ocaml in Dockerfile, use alpine directly
The use of FROM:returntocorp/ocaml:..., which are docker images
generated by ocaml-layer (https://github.com/semgrep/ocaml-layer/),
served us well in the past to speedup
CI. However, we now use GHA caching more and depot.dev
which removes the need for ocaml-layer. Worse, ocaml-layer
adds an extra layer of complexity that is annoying when
updating to more recent version of OCaml (e.g., 5.2).
For example I had trouble in the past in
https://github.com/semgrep/ocaml-layer/pull/46
or
https://github.com/semgrep/semgrep/pull/10213
to fix our ocaml-layer.
Moreover ocaml-layer relies on circleCI, which we don't
support anymore, so it's lots of extra work on ocaml-layer
that we don't really need.

Let's remove this intermediate.

test plan:
make build-docker