Change aliengrep metavariable kind to align with spacegrep (#10223)
Aliengrep now uses the same function as Spacegrep to extract a
metavariable value (`Xpattern_matcher.mval_of_string`), resulting in an
int or string literal instead of a `Text` node.

This avoids a problem that was apparently known for the "generic mode" -
presumably just Spacegrep at the time - since there was a comment about
it. I removed the comment because it no longer applies. I don't fully
understand what's going on with this `opt_xlang` value in
`Metavariable_pattern.ml`: why it is `None` for both Spacegrep and
Aliengrep?

Anyway, now both generic modes do the same thing so it's a little
simpler. The relevant change is just one line. In the course of the
investigation, I changed the metavariable type used by Aliengrep from a
tuple to a record to minimize future confusion.

Fixes https://github.com/semgrep/semgrep/issues/10222