feat: add argument to specify the number of jobs being used for interfile
In combination with changes to semgrep-proprietary, this produces a speedup for
interfile runs that use both intrafile and interfile rules. Instead of running
both the intrafile and interfile rules with `-j 1`, it runs first the intrafile
rules with the default number of jobs, and then the interfile rules with a single
job. This conserves memory for the interfile rules while allowing the intrafile
rules to benefit from multiprocessing.

TODO: we can remove the supply chain hack that disables interfile when possible
to get multiprocessing

Test plan: see corresponding semgrep-proprietary PR