feat: use per-product App ignores
Previously we only consumed a single "global" ignore list from the App.
With this, we'll use the new `product_ignored_files` and only use the
deprecated `ignored_files` as a fallback if `product_ignored_files` is
not present. Existing plumbing makes it so that this merely requires a
swap out for the ignore_patterns property.