devops: add oss cli tests (semgrep/semgrep-proprietary#1524)
Runs the CLI tests from the OSS Repo with the artifact built from this
repo. I disable one test in the CI job which seems to have had a
regression or something. I need to investigate the difference in the
snapshots, but I will do that in a follow-up PR.


PR checklist:
- [x] the commit message does not contain sensitive information (if

synced from 6972a795f2778a014daaf787694a397d3c4d0be7