fix: Fall back when unable to infer type from svalue (semgrep/semgrep-proprietary#1616)
During matching, when determining the type of an expression to use when
evaluating a typed metavariable, we previously would infer a type from
an svalue if it existed. This makes sense: typically, a propagated
constant will have a more specific type than what was declared. However,
if we're unable to extract a valid type from from the propagated
constant (for example if it is null), then we should fall back to the
resolved type for the name.

Test plan: Automated tests

synced from Pro 9e7e94aa0f9ef377401ce8ea14eda7536785dc81