chore: add logging to osemgrep sarif formatter (semgrep/semgrep-proprietary#1626)
Previously, when osemgrep vs pysemgrep has a mismatch for sarif, it will
only log the diff in `--debug`.

This PR adds (a limited amount of) diffs in `--verbose` so it's not too
verbose, but still helpful in situations like CI.

Tested manually with `--config r/all` on the semgrep repo. The
`--verbose` flag caps the diff, so it's not too verbose. The `--debug`
flag shows the whole diff as expected.

Also tested when the diff is smaller than the cap by faking a version
mismatch (modifying `__VERSION__`), and `--verbose` still logs
correctly.

synced from Pro 717a2eefe7a26d753c735d04bab01d5d984c4595