feat(rule-syntax): multiple fixes (semgrep/semgrep-proprietary#1541)
This is a stacked diff whose parent is #1539 

## What:
This PR makes it so that the new rule syntax permits you to write rules
which associate fixes to each pattern, as opposed to the entire rule.

## Why:
This will make it such that a single rule can potentially give many
possible fixes. This is better for rules where the fix is dependent upon
where in the rule the autofix is coming from.

## How:
Added a field `fix_text` to `Pattern_match.t` which is populated upon
`evaluate_formula`.

You may now write rules of the form:
```
rules:
  - id: rulename
    severity: ERROR
    languages: [python]
    message: |
      This rule has multiple fixes:
    match:
      any:
        - pattern: foo
          fix: bar
        - pattern: bar
          fix: foo
```

## Test plan:
Added a test to the `rules_v2` directory. Now, you can specify a
`.fixed` file with any test which has fixes.

PR checklist:
- [x] the commit message does not contain sensitive information (if

synced from Pro 7d8992b5becf4ddb0276a006690d16145474709b