fix: Improve JS/TS hardcoded propagators with callbacks (semgrep/semgrep-proprietary#1677)
See the difference between these two examples:
* https://semgrep.dev/playground/s/QrZq4
* https://semgrep.dev/playground/s/3qw0r

Presumably, `x => y` is desugared to `x => { return y; }`. Therefore,
the hardcoded propagators only match if there is a return value. This
was clearly unintended. For things like `.map`, this is not likely to be
consequential because the whole point is that there is a return value.
However, `forEach` typically would not be used with a return value
unless the curly braces were omitted only for aesthetic purposes.

Test plan: Automated tests

synced from Pro 78a7d392ecf6233c9477310ee359250fa269c305