Ignore rules that don't satisfy min-version or max-version in Python (semgrep/semgrep-proprietary#1627)
This is a first stab at honoring min-version/max-version without running
into a YAML validation error.

This PR makes pysemgrep completely ignore rules that don't satisfy the
version constraints. It solves the problem of supporting `severity:
CRITICAL` but as a result we no longer show a warning when a rule is
skipped. If this is acceptable, we can merge this. Otherwise, more work
is needed to produce these warnings like we do on the OCaml side.

Removing the JSON Schema validation completely doesn't work because some
fields of the rule are consulted on the Python side and if they're
broken, we get ugly errors. This is the case for the `severity` field:
it needs to be well-formed for the Python code to work and the JSON
Schema validation is a good protection against this.

Current PR status: I need to check the e2e tests that are failing for me
locally, and add a changelog entry.

synced from Pro 5228636feff81125e41a40269b6b2ffc46f6bd0e