CFG_build: Fix CFG for `try` without `finally` (semgrep/semgrep-proprietary#1695)
If there is no "finally" clause, then uncaught exceptions within the
"try" body should just be handled like if they were raised outside of
the "try".

Plus I fixed the throw destination in case that we are not in a nested
`try`, so that we throw to the exit node of the function.

Closes FLOW-78

test plan:
make test # added new test

synced from Pro 876f04fefee92c3fa70b3c907872fe64ea32ea07