refactor: remove AST_generic_equals (semgrep/semgrep-proprietary#1735)
Currently this module provides for two notions of equality, "semantic"
and "syntactic" (previously we had a third, "referential", too, but this
has since been removed), which can each be used to compare values of the
generic AST. The way this is done is via a ref which sets the type of
equality being used. Since this results in global state controlling this
equality it will need to be changed for OCaml multicore/5.x.

In fact, we only use this module for the syntactic equality in the
current draft version of steps mode, where it may not be the best choice
anyway (per Emma). Therefore, we remove this to simplify the code and
eliminate complexity we don't need.

Co-authored-by: Emma Jin <emjin@users.noreply.github.com>

synced from Pro cc44aba800c6a5006a5a0c1da5e6acd83ce8f42c