feat: end-to-end stub for pattern when implementation (semgrep/semgrep-proprietary#1732)
This PR adds components to the code for `pattern-when` (now incorporated
as part of `comparison`) to work. The logic for it is not implemented in
this PR and the functions are stubbed.

Test plan: run `make test` and it should not break existing
functionalities.

synced from Pro bf4175c69253d277df7ea07f96e753582c6430b2