feat(rule syntax): embed matches into metavariables (semgrep/semgrep-proprietary#1730)
## What:
This PR makes it such that you can bind arbitrary matches into
metavariables.

## Why:
Metavariables enjoy a lot of features. We have many operators, such as
`metavariable-type`, `metavariable-pattern`, `metavariable-regex`, etc,
which let us condition matches upon whether metavariables match a
certain condition.

This, however, only applies to things we can actually bind to
metavariables. This is a subset of AST nodes, because metavariables can
usually only take identifier position (among a few other special cases),
and only match typically to expression nodes.

Suppose, for instance, we wanted to take all Python functions and add a
decorator to them. We would write the following pattern:
```
pattern: |
  def $FUNC(...):
    ...
```

But what should we write for the fix?
```
fix: |
  @required_decorator
  def $FUNC(...):
    ...
```
This doesn't work, because we've lost the contents of the function body
and function arguments!

We can try something slightly different:
```
pattern: |
  def $FUNC($...ARGS):
    $...BODY
fix: | 
  @required_decorator
  def $FUNC($...ARGS):
    $...BODY
```
but this is really hacky, and won't work for all such statements we
would like to match, for instance if we cannot capture all the
nonterminals of the node with metavariables or metavariable ellipses.

Instead, it would be nice if we could just give a name to the match
produced on the function, and use it in the fix.
That's what this PR does.

## What:
This PR introduces the `as` syntax, which modifies an existing pattern
to bind that pattern's matches into a metavariable of a certain name.
For instance, in the example given above, we can write:
```
pattern: |
  def $FUNC(...):
    ...
as: $FUNCTION_DEF
fix: |
  @required_decorator
  $FUNCTION_DEF
```

This is simple, requires no knowledge of things like metavariable
ellipses, and generalizes to matching any kind of node. Now, if
something is possible to be a match, we can use tools like
`metavariable-regex` on it. This is greatly helpful in the Autofix case,
where we are quite sensitive to the range of the match.

## Test plan:
Automated tests.

Closes SAF-651

synced from Pro aaaf3ff495ffc166e7611b1d828b23cafec06ddd