refactor: Hide representation of pre/post processor set (semgrep/semgrep-proprietary#1685)
I'd like to change the representation to make error handling easier.
That's easier to do if the representation is hidden.

As part of this, I've also made it clear that the hook is only to be
used directly for testing purposes.

synced from Pro 18aa170aaee426ecc7aa3b05d79e0d094c04e6d6